chore(search): drop the Bleve auto-upgrade backend
The store-native FTS adapter is the production text backend, so the
BM25 -> Bleve auto-upgrade never fires: its trigger is guarded on the
active backend not already being the SymbolSearcher adapter, which
every non-test store construction satisfies.

Removes BleveBackend and the indexer's upgrade machinery (snapshot,
goroutine spawn, once-gating fields), leaving BM25 as the fallback for
stores without native search. Drops bleve and 25 satellite modules.
