+++ internal/config/config.go
@@ -264,15 +264,14 @@ type SemanticConfig struct {
 	SkipEmbed []SkipEmbedRule `mapstructure:"skip_embed" yaml:"skip_embed,omitempty"`
 
 	// SkipSearch lists (language, kind) combinations that should be
-	// kept in the graph but excluded from the text search index
-	// (BM25/Bleve). Same shape as SkipEmbed but targets a different
-	// index. The motivating case: a big monorepo with ~135k JSON
-	// `variable` nodes (package.json keys, tsconfig entries, etc.)
-	// pushed total symbol count over search.AutoThreshold and
-	// triggered an auto-upgrade from BM25 (~900 B/doc) to Bleve
-	// (~32 KiB/doc). Those config-key nodes aren't useful search
-	// targets — users who want to find them by name still can via
-	// graph queries. Defaults are a superset of SkipEmbed because
+	// kept in the graph but excluded from the text search index.
+	// Same shape as SkipEmbed but targets a different index. The
+	// motivating case: a big monorepo with ~135k JSON `variable`
+	// nodes (package.json keys, tsconfig entries, etc.) inflated the
+	// symbol corpus without ever being searched by name. Those
+	// config-key nodes aren't useful search targets — users who want
+	// to find them by name still can via graph queries. Defaults are
+	// a superset of SkipEmbed because
 	// anything that isn't worth embedding usually isn't worth
 	// full-text-indexing either. See DefaultSkipSearch.
 	SkipSearch []SkipEmbedRule `mapstructure:"skip_search" yaml:"skip_search,omitempty"`
@@ -343,9 +342,9 @@ func DefaultSkipEmbed() []SkipEmbedRule {
 
 // DefaultSkipSearch returns the baseline (language, kind) pairs that
 // are kept out of the text search index. Superset of DefaultSkipEmbed:
-// if a node isn't worth a vector slot it generally isn't worth a BM25/
-// Bleve slot either, and on big monorepos these config-key nodes are
-// what pushes the backend into its Bleve auto-upgrade (~32 KiB/doc).
+// if a node isn't worth a vector slot it generally isn't worth a
+// full-text slot either, and on big monorepos these config-key nodes
+// dominate the corpus without ever being searched by name.
 // JSON is the heaviest of the additions — tsconfig / package.json /
 // lockfile keys alone can account for >100k variable nodes.
 func DefaultSkipSearch() []SkipEmbedRule {
